+++ src/main/java/de/samply/teiler/utils/Ping.java
@@ -60,8 +60,11 @@ public class Ping {
     private boolean openConnectionAndPing(String url) {
         try (CloseableHttpUrlConnection httpUrlConnection = new CloseableHttpUrlConnection(url)) {
             logger.info("Ping to " + url + "...");
-            return ping(httpUrlConnection);
+            boolean result = ping(httpUrlConnection);
+            logger.info("[" + ((result) ? "reachable" : "unreachable") + "]");
+            return result;
         } catch (IOException e) {
+            logger.info("[unreachable]");
             return false;
         }
     }